+++ private/tpl/group/limited.php
@@ -15,14 +15,13 @@ if (!defined('RAVEN_VIEW_RENDER_CONTEXT')) {
 <section>
 {if group:name}
     <h1 class="mb-3">Group: {group:name}</h1>
-{/if}
-{if not group:name}
+{else}
     <h1 class="mb-3">Group</h1>
 {/if}
 
 {if group:slug}
     <p class="text-muted">Slug: {group:slug}</p>
 {/if}
-    <p class="text-muted">{group:member_count} Users</p>
+    <p class="text-muted">{group:count} Users</p>
     <p class="text-muted">Limited group view.</p>
 </section>